requirements.txt
pyusb
paho-mqtt
mido
